refactor: Add debug page
Add debug page for displaying extra information during deevelopment.

This page can link to credential issuers, GPG45 calculations, or any
other information that we might find useful.

In development mode this page will be used instead of automatically
redirecting to the required credential issuer.

This page is not intended to be used in production.