BAU: Update signed JWT base64 regex
The previous version of the base64 regex was being flagged by Sonarcloud
as being susceptible to DOS due to backtracking. Whilst we control the
input to this regex so this shouldn't be too much of a concern, it would
be good to clear this up just in case we fall foul of it
unintentionally.

This version checks that we have three groups of base64url characters (+
and / have been replaced with - and _). The first two groups must be
suffixed with a ".". The shared attribute JWTs we're returning from
core-back should never deviate from this pattern.