PYIC-875: Return state in auth response
The Oauth spec states that if a value for state is sent in the
authentication request, we should return in in the response. See here:
https://datatracker.ietf.org/doc/html/rfc6749#section-4.1.2

Core-back has been updated to return the state value supplied in the
auth request. The state value is not mandatory so may not appear in the
client data received from core-back. This change will add it as a query
param to the redirect URL if it's provided.

The auth teams client validates that the state received in the response
is the same as that sent in the initial auth request. We need to return
it to allow this to happen.