PYIC-1028 Deploy Core Front to ECS Fargate
This template creates the necessary components to deploy Core Front to
ECS Fargate behind an API Gateway. Core Front ECS Service runs in a
private subnet serviced by a private Application Load Balancer. A public
API Gateway is linked to the private ALB via VPCLink (see
https://github.com/alphagov/digital-identity-architecture/blob/main/adr/0025-relaxing-serverless-constraints.md#all-end-user-facing-http-elements-need-to-be-behind-api-gateway).

Core Front will continue to call Core Back's API Gateway via the public
internet though this change to bring Core Front into Amazon makes it
possible for that communication to be done via a VPC endpoint in future
and keep the traffic within Amazon's network.

The VPC, private Subnets and NAT gateways are created in a separate
stack and the VPC and subnet ids are provided as parameters.