BAU Do not Output ALB Access Logs in Developer Environments
The ALB access logs bucket makes it harder to destroy the core-front
stacks because the ALB access logs bucket cannot be deleted by CFN once
it has objects in it. There is little value in ALB access logs in the
developer environments and if absolutely necessary this can be enabled
on their environment as and when needed.