BAU: Spike into skipping page redirect
This is an experiment to see if we can render pages directly when we
receive a page journey response from the backend.

This renders pages at the `/ipv/journey/next` page, rather than
redirecting the user again to a `/journeyPage?pageid=...` URL.

It saves a round trip for the user but might be confusing rendering
different pages on the `/next` page.

This is not tested and may be (probably is) handling errors incorrectly.