BAU: Add validate endpoint
We have a new lambda for validating the evidence in a VC to ensure a
user can carry on with their journey. We need to allow this action in
the frontend.

The list was getting a bit long so have refactored to use regex's which
are a little more concise.