BAU: Set trust proxy on express app
Currently the frontends session cookies are not being set with the
Secure attribute, which they probably should be.

The hmpo-app framework we're using is setting the session cookie option
to `secure: 'auto'` which means it checks the `x-forwarded-proto` header
to determine if the original request was https (the TLS gets terminated
at the API gateway for us)

Express won't trust or use the `X-Forwarded` headers unless you tell it
to: https://expressjs.com/en/guide/behind-proxies.html

This sets up the app to trust at most a reverse proxy two hops away.
API Gateway -> ALB -> Express