PYIC-1281 Checkov Fixes
Cleans up the following Checkov recommendations. There are more to do.

Enable ALB Access Logging: This is purposefully disabled for developer
environments to make them easier to manage. It is enabled for other
environments

TLS Listeners on ALB: Suppressed this warning because this is not
possible on a private ALB behind an API Gateway due to certificate
generation. This has been discussed and agreed as acceptable at the
security working group.

S3 Access Logging: Suppressed because this is the bucket which holds the
S3 access logs and AWS recommend that a bucket does not send its own
access logs to itself.

ECS Container Insights: Enabled because this will be useful for
observability.