PYIC-1281 Fix Bucket Policy Dependency
The access logs bucket policy must be applied before the ALB is created
which needs access to the bucket to send its access logs to. Previously
we were only creating the access logs bucket and its policy in
non-development environments. Unfortunately the CFN `DependsOn` does not
support `!If` conditions. Therefore it is necessary to create the bucket
and its policy in all environments, however we can continue to only send
ALB access logs to it in non-development environments.