BAU: Validate that page requests are from a set list
We need to avoid the possiblity of a traversal attack on our input pageid. We are likely to rework how we handle page loading, but we need to put a fix in place quickly.

This change adds a set of case statements for all the valid pages, redering those value and defaulting to an error page.