PYIC-1511 Added new deploy_dev script
This works in a silightly different fashion it keeps the config-repo and
instead runs a git pull on each invocation to keep itself constantly
updated.

I have a fast connection so for me

deploy_dev first run range 3.2 - 3.8s consequent runs 1.4s - 1.6s

deploy_to_dev_env.sh always around 1.7 - 2s

for me not much difference (I prefer the deploy_to_dev_env as it is
cleaner imho), however for those on a slower connection maybe the
deploy_dev method will be better initially it will be a bit slower
but consequent runs should be faster.

I have uploaded both methods to see which one preforms better, for
others or which is preferred.

	new file:   deploy_dev.sh