PYIC-1484: Adjust app log
We've configured the app logs to log the ipvSessionId by pulling it from
the session. To do this we need to pass the request object to any
application logs we generate. For example:

        `logger.info('hello world', {req})`

Once the logger has the request object, it just needs to call,
`session.ipvSessionId` on it, without the `req` prefix we originally
included in the config. This has been tested in the passport front.