BAU: Refactor handleJourneyResponse method
This splits the `handleJourneyResponse` method in two. This allows the
routing logic for backend responses to be used without having to call
the backend.

This has been done in preparation for the credential-issuer middleware
to be able to receive more than a journey response. A page response, for
example, in the case that there is an error in one of the CRI response
lambdas.

This is needed as there is a piece of work to run the CRI response
lambdas in a step function, rather than bouncing back to the frontend.
The step function may return a page response, and so the
credential-issuer middleware has to be able to handle it.