BAU: Use new `/cri/callback` backend endpoint
The new backend endpoint hooks into the CRI return step function. This
is a more appropriate endpoint that `/cri/validate-callback`.

This also switches the request content type to `application/json`. We're
dealing with JSON in the backend lambdas so it makes sense to keep
things simple. This change will also allow us to directly deserialize
the incoming data in the validate-callback lambda, rather than using an
input stream and having to read it.