BAU: Don't use logging middleware for healthcheck
The logging middleware adds a custom message for each message received
and each response. This is very noisy when you start pinging the app
with healthchecks every 30 seconds.

This moves the middleware from being registered on the entire app to
being registered on the router for the usual endpoints.

It also creates a brand new router for the healthcheck endpoint. This
router doesn't have the logging middleware applied to it so we can avoid
the spam.

IMPORTANT - the order that the routers are registered on the app with
`app.use()` is important. The healthcheck router needs registering
first, or the logging middleware will cascade down on to it. Very
annoying.