Ensure session is saved before redirect
When running locally, the async nature of the code meant that the
redirect was issued and followed by the client before the session had
finished saving. This happened about 50% of the time.

Moving the redirect to a callback from calling session.save prevents
this.

This isn't production ready as it has no error handling, but this might
be worth thinking about changing.