BAU: Save session before internal redirect
While getting core-back and core-front to run locally it was noticed
that it's possible for core-front to issue a redirect, and it be
followed, before the session has finished writing. This resulted in
weird errors where the expected page didn't match the request page.

Whilst we've not knowingly seen this happen in production, probably due
to the longer request times involved, it's a possibility and we should
address it.

This updates core front to issue the redirect in the callback function
of the session.save() method. This will mean the redirect will only be
issued once the session is successfully stored.

This has only been applied to internal redirects - redirects that take
the user to a different page on core front. We don't need to worry about
this issue if we're sending the user off to do something elsewhere.