PYIC-6395: Fix page-ip-reuse ordering and content
This updates the ordering of the rows on the page-ipv-reuse page - the
DOB should have been below the address.

This also ensures that we only show the current address for the COI
version - previously we would have rendered previous addresses.

It also fixes the content for "Full name", in both English and Welsh.