PYIC-6833: Markup COI start page details with H3
The "If your details are wrong" dropdown in the details component of the
reuse page needed to be marked as an H3 to fit properly with the other
headings on the page, for accessibility reasons.

To do this we needed to stop using the nunjucks macro and instead
include the html for the component directly. The macro does allow you to
pass HTML in for the dropdown, but it all gets put inside a `span`
element, and an h3 shouldn't be inside a span.

We also need to override the font size and weight from the h3. This is
done tith the override classes made available by the design system.