PYIC-7613: Update api gateway 5XX alarm
This updates the existing alarm for the percentage of 5XX errors from
the frontend API gateway.

The biggest change is the API ID being targeted - previously it was
connected to an unused (should we remove it?), http API gateway. It was
also set to alert if error rates were 1%. This fixes it to 10%.

It also removes the check for if the number of errors is a certain count
- as we're using a percentage this is implicit in the lower limit on
  number of invocations.