PYIC-7869: Render service-unavailable page on 503
The overload protection lib doesn't seem to be configurable to specify a
template or html page to return when it kicks in. We can however set it
to hand control back to the framework and handle a 503 ourselves.

The downside of this is that it will require the app to do work when
it's already overloaded.