BAU: Fix overload-protection logging
I noticed that there we no logs being generated when overload-protection
kicked in.
Specifying the logging config as a string value requires that req.log
exists. We haven't assigned that yet in the middleware stack, at the
point overload protection is run. So we got no logs.

We can instead set the config for logging to a function which is
responsible for the logging.