SPT-1693: Allow access to published keys KMS key
The published keys bucket is migrating from using SSE-S3 to KMS-S3
encryption. This change adds the ability for the
RestDidDocumentResourceS3Role to access the encryption key in the
published bucket account. A matching change is being made in the account
to allow IPV Core to perform a Decrypt operation using the key.

Co-authored-by: Chris Clayson <chris.clayson@digital.cabinet-office.gov.uk>