INC0016299 Set headersTimeout value greater than keepAliveTimeout
We've been seeing a handful of 502s from our ALB. The root cause is unclear but I've discovered a Node issue
which suggests setting the Node headersTimeout value greater than the keepAliveTimeout to prevent a possible
ALB 502 race condition.

Node enforces a default headersTimeout of 60000ms. We manually set keepAliveTimeout to 65000ms to handle ALB
idle timeouts. This creates a 5-second window (60s-65s) where the ALB considers a connection valid but Node
has internally marked it for destruction. If a request arrives in this window, Node resets the connection
causing a 502.

See https://github.com/nodejs/node/issues/27363.